refactor(chat): extract getArtistSegments lib helper from hook
Following project pattern (fetchArtists, deleteArtist) — network call lives
in lib/artists/*, the hook just wires Privy auth + react-query.
